requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
openai==1.107.0
python-dotenv==1.1.1
pydantic==2.11.7
requests==2.32.5
streamlit==1.49.1
